moved the function in the class
